Fixed issues:  - using CSV format to copy-paste data from and to editor  - context column visibility real-time change  - reference lookuper thread sync  - paste - folder  - toolbar remove - embedded resources  - sync case sensitivity  - global translate - bing  - move to resources on aspx - webapplication  - import statement in aspx  - website files code model  - 'extern' methods fix  - default custom criteria  - \0 escape  - error on unknown attribute  - properties in VB  - error on solution explorer closed  - referencing using My feature in VB

+++ VisualLocalizer/VisualLocalizer/Commands/Inline/AbstractBatchReferenceProcessor.cs
@@ -75,7 +75,7 @@ namespace VisualLocalizer.Commands {
                         string text = GetReplaceString(resultItem); 
                         
                         string path = resultItem.SourceItem.GetFullPath();
-                        if (RDTManager.IsFileOpen(path)) { // file is open
+                        if (RDTManager.IsFileOpen(path) && RDTManager.IsFileVisible(path)) { // file is open
                             if (!buffersCache.ContainsKey(path)) { // file's buffer is not yet loaded
                                 // load buffer
                                 IVsTextLines textLines = DocumentViewsManager.GetTextLinesForFile(path, false);
@@ -120,7 +120,14 @@ namespace VisualLocalizer.Commands {
             }
 
             foreach (var pair in filesCache) {
-                File.WriteAllText(pair.Key, pair.Value.ToString());
+                if (RDTManager.IsFileOpen(pair.Key)) {
+                    RDTManager.SetIgnoreFileChanges(pair.Key, true);
+                    File.WriteAllText(pair.Key, pair.Value.ToString());
+                    RDTManager.SetIgnoreFileChanges(pair.Key, false);
+                    RDTManager.SilentlyReloadFile(pair.Key);
+                } else {
+                    File.WriteAllText(pair.Key, pair.Value.ToString());
+                }
             }
             if (errorRows > 0) throw new Exception("Error occured while processing some rows - see Output window for details."); 
         }
